Move properties builder to the form builder
