+++ flow-model-generator/modules/flow-model-generator-core/src/main/java/io/ballerina/flowmodelgenerator/core/model/Branch.java
@@ -106,7 +106,7 @@ public record Branch(String label, BranchKind kind, Codedata codedata, Repeatabl
         private Repeatable repeatable;
 
         protected Codedata.Builder<Builder> codedataBuilder;
-        protected NodeBuilder.PropertiesBuilder<Builder> propertiesBuilder;
+        protected FormBuilder<Builder> formBuilder;
         private SemanticModel semanticModel;
         private DiagnosticHandler diagnosticHandler;
         private ModuleDescriptor moduleDescriptor;
@@ -155,12 +155,12 @@ public record Branch(String label, BranchKind kind, Codedata codedata, Repeatabl
             return this;
         }
 
-        public NodeBuilder.PropertiesBuilder<Builder> properties() {
-            if (this.propertiesBuilder == null) {
-                this.propertiesBuilder =
-                        new NodeBuilder.PropertiesBuilder<>(semanticModel, diagnosticHandler, moduleDescriptor, this);
+        public FormBuilder<Builder> properties() {
+            if (this.formBuilder == null) {
+                this.formBuilder =
+                        new FormBuilder<>(semanticModel, diagnosticHandler, moduleDescriptor, this);
             }
-            return this.propertiesBuilder;
+            return this.formBuilder;
         }
 
         public Codedata.Builder<Builder> codedata() {
@@ -172,7 +172,7 @@ public record Branch(String label, BranchKind kind, Codedata codedata, Repeatabl
 
         public Branch build() {
             return new Branch(label, kind, codedataBuilder == null ? null : codedataBuilder.build(),
-                    repeatable, propertiesBuilder == null ? null : propertiesBuilder.build(), children);
+                    repeatable, formBuilder == null ? null : formBuilder.build(), children);
         }
     }
 }
