+++ flow-model-generator/modules/flow-model-generator-ls-extension/src/test/resources/suggested_flow_model/config/if4.json
@@ -223,7 +223,8 @@
               {
                 "id": "42941",
                 "metadata": {
-                  "label": "print"
+                  "label": "print",
+                  "description": ""
                 },
                 "codedata": {
                   "node": "FUNCTION_CALL",
@@ -306,7 +307,8 @@
               {
                 "id": "44921",
                 "metadata": {
-                  "label": "print"
+                  "label": "print",
+                  "description": ""
                 },
                 "codedata": {
                   "node": "FUNCTION_CALL",
@@ -424,7 +426,8 @@
                       {
                         "id": "46042",
                         "metadata": {
-                          "label": "print"
+                          "label": "print",
+                          "description": ""
                         },
                         "codedata": {
                           "node": "FUNCTION_CALL",
@@ -507,7 +510,8 @@
                       {
                         "id": "48030",
                         "metadata": {
-                          "label": "print"
+                          "label": "print",
+                          "description": ""
                         },
                         "codedata": {
                           "node": "FUNCTION_CALL",
