+++ flow-model-generator/modules/flow-model-generator-ls-extension/src/test/resources/suggested_flow_model/config/if7.json
@@ -287,7 +287,8 @@
               {
                 "id": "43072",
                 "metadata": {
-                  "label": "print"
+                  "label": "print",
+                  "description": ""
                 },
                 "codedata": {
                   "node": "FUNCTION_CALL",
@@ -474,7 +475,8 @@
               {
                 "id": "49885",
                 "metadata": {
-                  "label": "print"
+                  "label": "print",
+                  "description": ""
                 },
                 "codedata": {
                   "node": "FUNCTION_CALL",
@@ -556,7 +558,8 @@
               {
                 "id": "51865",
                 "metadata": {
-                  "label": "print"
+                  "label": "print",
+                  "description": ""
                 },
                 "codedata": {
                   "node": "FUNCTION_CALL",
